validate and sanitized login user data
