+++ routes/authRoutes.js
@@ -2,13 +2,13 @@ const { Router } = require('express');
 const router = Router();
 
 const { register, login, me } = require('../services/auth.service');
-const {validateUser} = require('../utils/verifiedRegister')
+const {validateUser, valideLogin} = require('../utils/verifiedData')
 
 function authRoutes(app) {
   app.use('/auth', router);
 
   router.post('/register', validateUser, register);
-  router.post('/login', login);
+  router.post('/login', valideLogin, login);
   router.post('/me', me);
 
 }
